Rename package stringHandlers to io
The old name felt too long
