+++ src/main/java/yapper/instructions/InstructionHandler.java
@@ -3,8 +3,8 @@ package yapper.instructions;
 import yapper.TaskManager;
 import yapper.Yapper;
 import yapper.exceptions.ErrorHandler;
-import yapper.stringHandlers.InputStringHandler;
-import yapper.stringHandlers.OutputStringHandler;
+import yapper.io.InputStringHandler;
+import yapper.io.OutputStringHandler;
 import yapper.tasks.Task;
 import yapper.tasks.Deadline;
 import yapper.tasks.Event;
