+++ src/main/java/yapper/io/OutputStringHandler.java
@@ -1,4 +1,4 @@
-package yapper.stringHandlers;
+package yapper.io;
 
 import yapper.exceptions.ErrorHandler;
 import yapper.exceptions.YapperException;
