+++ src/main/java/yapper/io/StringStorage.java
@@ -1,4 +1,4 @@
-package yapper.stringHandlers;
+package yapper.io;
 
 // Contains strings for the Yapper program
 public class StringStorage {
